+++ MadnessTests/ParserTests.swift
@@ -57,43 +57,4 @@ final class ParserTests: XCTestCase {
 	func testAnyParsesAnySingleCharacter() {
 		assertTree(any, "🔥", ==, "🔥")
 	}
-
-
-	// MARK: - Nonterminals
-
-	// MARK: Ignoring
-
-	let ignored = ignore("x")
-
-	func testIgnoredInputDoesNotGetConcatenatedAtLeft() {
-		assertTree(ignored ++ %"y", "xy", ==, "y")
-	}
-
-	func testIgnoredInputDoesNotGetConcatenatedAtRight() {
-		assertTree(%"y" ++ ignored, "yx", ==, "y")
-	}
-
-	func testIgnoringDistributesOverConcatenation() {
-		assertAdvancedBy(ignored ++ ignored, "xx", 2)
-	}
-
-	func testIgnoredInputIsDroppedFromAlternationsAtLeft() {
-		assertTree(ignored | %"y", "y", ==, "y")
-	}
-
-	func testIgnoredInputIsDroppedFromAlternationsAtRight() {
-		assertTree(%"y" | ignored, "y", ==, "y")
-	}
-
-	func testIgnoringDistributesOverAlternation() {
-		assertMatched(ignored | ignored, "x")
-	}
-
-	func testRepeatedIgnoredEmptyParsesAreDropped() {
-		assertTree(ignored* ++ %"y", "y", ==, "y")
-	}
-
-	func testRepeatedIgnoredParsesAreDropped() {
-		assertTree(ignored* ++ %"y", "xxy", ==, "y")
-	}
 }
